Added new design for profile, settings, for creating events and some fixes in event page, create bottom navigation for mobile. Optimized application for mobile devices

+++ src/hooks/useChangeFormatDate.ts
@@ -9,7 +9,7 @@ export const useChangeFormatDate = (params: IChangeFormatDate) => {
     const dateObj = new Date(dateValue);
 
     // Use toLocaleDateString to format the date
-    const options: any = { day: 'numeric', month: 'long' };
+    const options: any = { day: 'numeric', month: 'short' };
     if (language) {
         options.localeMatcher = 'lookup';
         options.locale = language;
